Only cache the pip cache
Unfortunately, due to https://github.com/wntrblm/nox/issues/735,
caching the nox virtual environments causes random intermittant
GitHub Actions failures. Until that issue is fixed, only cache the
pip cache and remove the settings for a cache prefix and for the
dependency glob to decide when to restore the nox virtual environment
cache.