Cache by Python location instead of version
python-version only contains the minor version, but the path to
Python includes the patch version. We seem to be sometimes getting
Python 3.12.0 and sometimes getting Python 3.12.1, and I suspect
this is what may be causing the caching issues. Try caching by
Python location using a cache key syntax similar to that of
pre-commit to see if that addresses the problem.